move classes into distinct file
